fix: Исправил замечания
