+++ src/main/java/ru/practicum/shareit/user/model/User.java
@@ -2,9 +2,6 @@ package ru.practicum.shareit.user.model;
 
 import lombok.*;
 
-/**
- * TODO Sprint add-controllers.
- */
 @Getter
 @Setter
 @AllArgsConstructor
